Installed Devise views for overriding and styled for Bulma
The default Devise views do not have a lot of unique styling. Since
we are using the Bulma framework we need to customize these views
to look nice with Bulma. Installed a new layout specifically for
Devise actions to unify the styling across Devise controllers while
maintaining a separate site layout.